refactor: Member Api Request 관련된 모델끼리 Kotlin File로 응집
