+++ src/main/java/com/koddy/server/member/presentation/request/CompleteProfileRequest.kt
@@ -1,11 +1,13 @@
 package com.koddy.server.member.presentation.request
 
+import com.koddy.server.member.application.usecase.command.CompleteMenteeProfileCommand
 import com.koddy.server.member.application.usecase.command.CompleteMentorProfileCommand
+import com.koddy.server.member.presentation.request.model.MentoringPeriodRequestModel
 
 data class CompleteMentorProfileRequest(
     val introduction: String?,
     val profileImageUrl: String?,
-    val period: MentoringPeriodRequest?,
+    val period: MentoringPeriodRequestModel?,
     val schedules: List<MentorScheduleRequest> = emptyList(),
 ) {
     fun toCommand(mentorId: Long): CompleteMentorProfileCommand =
@@ -15,5 +17,17 @@ data class CompleteMentorProfileRequest(
             profileImageUrl,
             period?.toPeriod(),
             schedules.map { it.toTimeline() },
+        )
+}
+
+data class CompleteMenteeProfileRequest(
+    val introduction: String?,
+    val profileImageUrl: String?,
+) {
+    fun toCommand(menteeId: Long): CompleteMenteeProfileCommand =
+        CompleteMenteeProfileCommand(
+            menteeId,
+            introduction,
+            profileImageUrl,
         )
 }
