+++ src/main/java/com/koddy/server/member/presentation/request/LookAroundMembersByConditionRequest.kt
@@ -1,13 +1,36 @@
 package com.koddy.server.member.presentation.request
 
 import com.koddy.server.global.utils.FilteringConverter
-import com.koddy.server.member.application.usecase.query.GetMenteesByCondition
+import com.koddy.server.member.application.usecase.query.LookAroundMenteesByConditionQuery
+import com.koddy.server.member.application.usecase.query.LookAroundMentorsByConditionQuery
 import com.koddy.server.member.domain.model.Language
 import com.koddy.server.member.domain.model.Nationality
 import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotNull
 
-data class GetMenteesByConditionRequest(
+data class LookAroundMentorsByConditionRequest(
+    val languages: String?,
+
+    @field:NotNull(message = "페이지 번호는 필수입니다.")
+    @field:Min(value = 1, message = "페이지는 1부터 시작입니다.")
+    val page: Int,
+) {
+    fun toQuery(): LookAroundMentorsByConditionQuery {
+        return LookAroundMentorsByConditionQuery(
+            convertToLanguageCategory(),
+            page,
+        )
+    }
+
+    private fun convertToLanguageCategory(): List<Language.Category> {
+        if (languages.isNullOrBlank()) {
+            return listOf()
+        }
+        return FilteringConverter.convertToLanguage(languages)
+    }
+}
+
+data class LookAroundMenteesByConditionRequest(
     val nationalities: String?,
 
     val languages: String?,
@@ -16,8 +39,8 @@ data class GetMenteesByConditionRequest(
     @field:Min(value = 1, message = "페이지는 1부터 시작입니다.")
     val page: Int,
 ) {
-    fun toQuery(): GetMenteesByCondition {
-        return GetMenteesByCondition(
+    fun toQuery(): LookAroundMenteesByConditionQuery {
+        return LookAroundMenteesByConditionQuery(
             convertToNationality(),
             convertToLanguageCategory(),
             page,
