+++ src/main/java/com/koddy/server/member/presentation/request/model/MentoringPeriodRequestModel.kt
@@ -1,9 +1,9 @@
-package com.koddy.server.member.presentation.request
+package com.koddy.server.member.presentation.request.model
 
 import com.koddy.server.member.domain.model.mentor.MentoringPeriod
 import java.time.LocalDate
 
-data class MentoringPeriodRequest(
+data class MentoringPeriodRequestModel(
     val startDate: LocalDate?,
     val endDate: LocalDate?,
 ) {
